requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-crispy-forms==1.7.2
django-rocket-engine==0.0.11
idna==2.8
Pillow==6.0.0
pytz==2018.9
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.1
virtualenv==16.4.3
